example of using different mixins
